requirements.txt
numpy==1.23.0
onnx==1.12.0
onnx_simplifier==0.3.10
onnxruntime==1.11.1
onnxsim==0.4.0
opencv_python==4.2.0.34
pycocotools==2.0.4
PyYAML==6.0
torchsummary==1.5.1
tqdm==4.59.0
